Update_max in go.c moved to separate loops
It turns out the four stocks that the flow_carbon function updates do not need their max updated after each flow of carbon.  This is because the MAX_DOC, MAX_POC, MAX_PHYTO, and MAX_WATERDECOMP are not actually used anywhere.  Since the other stocks are not updated in flow_carbon, it makes more sense to move update_max to outside the flow_carbon loop.

Also taking the opportunity to remove compiled code from my repo and gitignore files.